requirements.txt
python-dotenv==0.21.1
python-telegram-bot==13.15
openai==0.27.0
anthropic-cli==0.0.7